FUNDAMENTALLY CHANGED THE DOUBLY-LINKED-LIST DATA STRUCTURE!!!
A dll used to be a circular doubly linked list.  Unfortunately, this led to confusion as to the proper thing to do while traversing the list.

dlls are now doubly linked lists with a distinct head and tail.  confusion is done with!

All 104 tests pass under CMUCL including tests of some of the higher-level stuff!

This will constitute the 0.0.1 RC 8 release.