minor fix-ups.
added --count-rules command line option & associated code to allow rules and rulesets to be counted
for both the number of times they were used to check a message and how many times they found a matching
message.  The idea here is to allow the users to understand bottlenecks in their rulesets.