Added support for CLISP!
All but one test pass!  I believe that failure is due to a bug in CLISP!
This actually only affects the testing code; LoGS is trying to find a new file with a new inode  which is what syslog does.  CLISP just is having problems creating a new file with the same name as the old file and a new inode!

This will likely be tagged 0.0.4!