removed useless print statement when opening a file-follower with a position
fixed-up cl-cli to include the changes made to it for qgrokl, namely the addition of the help function (should be mapped to the --help option or some such by the user).  Instructed CMUCL to not complain about command line options its not aware of (because cl-cli handles those and CMUCL shouldn't care).